Create Why are you making peach spam bots
I have screenshots of the 400+ spam bot requests you sent me but no idea how to upload them. I recognize this is a bizarre means of contacting someone but I truly can find no other way to reach you and I would really love if you could delete these bots because peach is one of the last refuges free of spam online and it will be impossible for me to manually delete all of them, as I woke up to 50 more.